refactor(leases): extract useSharePnLDialog from SharePnLDialog.vue (#185)
Move the SharePnLDialog script verbatim into a co-located composable so the
SFC is template + one destructure, per the house Vue style. Behaviour-
preserving: all data derivation, the canvas render pipeline, and the
share/download paths move unchanged. `defineExpose({ show })` stays in the
SFC (it is a compiler macro) with the composable returning `show`; the
`dialog` ref is typed by the `{ show() }` surface it actually calls instead
of `typeof Dialog`, so the composable needs no web-components runtime import.
